Agent2_Outlet_Fit/requirements.txt
openai>=1.0.0
pypdf>=4.2.0
python-dotenv>=1.0.0
questionary>=2.0.0
